requirements.txt
hydra-core==1.3.2
lightrun==1.10.0
numpy==1.22.4
scipy==1.6.3
